+++ filefinder/tests/test_filecontainer.py
@@ -140,6 +140,28 @@ def test_filecontainer_search(example_df, example_fc):
     pd.testing.assert_frame_equal(result.df, expected)
 
 
+def test_filecontainer_concat(example_fc):
+
+    with pytest.raises(ValueError, match="Can only concatenate two FileContainers."):
+        example_fc.concat("not a FileContainer")
+
+    with pytest.raises(ValueError, match="FileContainers must have the same keys"):
+        different_keys_fc = FileContainer(example_fc.df.loc[:, ["model", "scen"]])
+        example_fc.concat(different_keys_fc)
+
+    result = example_fc.concat(example_fc, drop_duplicates=False)
+    expected = pd.concat([example_fc.df, example_fc.df])
+
+    pd.testing.assert_frame_equal(result.df, expected)
+    assert len(result) == 10
+
+    result = example_fc.concat(example_fc, drop_duplicates=True)
+    expected = example_fc
+
+    pd.testing.assert_frame_equal(result.df, expected.df)
+    assert len(result) == 5
+
+
 def test_fc_combine_by_key_deprecated(example_fc):
 
     with pytest.warns(FutureWarning, match="`combine_by_key` has been deprecated"):